refactor(tools): rename files and update async
