+++ tools/testSttFile.ts
@@ -1,4 +1,4 @@
-// 실행 명령어: pnpm tsx tools/test-stt-file.ts <파일경로>
+// 실행 명령어: pnpm tsx tools/testSttFile.ts <파일경로>
 // 예시: pnpm tsx tools/test-stt-file.ts my_voice.mp3
 
 import { SpeechService } from '../src/services/SpeechService';
